+++ src/Paramore.Brighter.ServiceActivator/Consumer.cs
@@ -86,11 +86,12 @@ namespace Paramore.Brighter.ServiceActivator
         /// <param name="subscriptionName">The name of the associated subscription.</param>
         /// <param name="channel">The channel.</param>
         /// <param name="messagePump">The message pump.</param>
-        public Consumer(ConsumerName name, SubscriptionName subscriptionName, IAmAChannel channel, IAmAMessagePump messagePump)
+        /// <param name="runAsync">Should we run this pipeline async?</param>
+        public Consumer(ConsumerName name, SubscriptionName subscriptionName, IAmAChannel channel, IAmAMessagePump messagePump, bool runAsync)
         {
             Name = name;
             SubscriptionName = subscriptionName;
-            Performer = new Performer(channel, messagePump);
+            Performer = new Performer(channel, messagePump, runAsync);
             State = ConsumerState.Shut;
         }